requirements.txt
azure-identity
azure-mgmt-resource
python-dotenv
openai
